added authentication and filtered contacts by user
-added authentication: email & password, and sign in with Google
-users now have their own contacts filtered by UID
-added protected routes
- changed contact 'description' in database to 'notes' and in UI

+++ server/index.js
@@ -6,17 +6,23 @@ const pool = require("./db");
 const multer = require("multer");
 const { CloudinaryStorage } = require("multer-storage-cloudinary");
 const cloudinary = require("cloudinary").v2;
+var admin = require('firebase-admin');
+var serviceAccount = require('./serviceAccountKey.json');
 
 // Middleware
+
 app.use(cors());
 app.use(express.json()); //req.body
+
+admin.initializeApp({
+    credential: admin.credential.cert(serviceAccount)
+});
 
 cloudinary.config({ 
     cloud_name: process.env.CLOUDINARY_NAME,
     api_key: process.env.CLOUDINARY_API_KEY, 
     api_secret: process.env.CLOUDINARY_API_SECRET
 });
-// Environment variables are usually set in a settings UI of a hosting platform when it comes to apps in production.
 
 const storage = new CloudinaryStorage({
     cloudinary: cloudinary,
@@ -29,14 +35,39 @@ const storage = new CloudinaryStorage({
 
 const parser = multer({ storage: storage });
 
+// verify Id token, attach decoded token to req
+const verifyToken = async (req, res, next) => {
+    const authorizationHeader = req.headers.authorization || '';
+    const components = authorizationHeader.split(' ');
+
+    if (components.length === 2) {
+        const type = components[0];
+        const token = components[1];
+
+        if (type === 'Bearer') {
+            try {
+                const decodedToken = await admin.auth().verifyIdToken(token);
+                req.user = decodedToken;
+                next();
+            } catch (e) {
+                res.status(403).json({ error: 'Unauthorized' });
+            }
+        } else {
+            res.status(400).json({ error: 'Invalid token' });
+        }
+    } else {
+        res.status(400).json({ error: 'Invalid token' });
+    }
+};
+
 
 // ======== ROUTES ======== //
 
 // Create a contact
 
-app.post("/contacts", parser.single("photo"), async (req, res) => {
+app.post("/contacts", verifyToken, parser.single("photo"), async (req, res) => {
     try {
-        const { first_name, last_name, email, phone, address1, address2, city, state, zip, categories, description } = req.body;
+        const { first_name, last_name, email, phone, address1, address2, city, state, zip, categories, notes } = req.body;
 
         // Check for required values
         if (!first_name) {
@@ -55,7 +86,9 @@ app.post("/contacts", parser.single("photo"), async (req, res) => {
             photo_upload_time = new Date();
         }
 
-        const newContact = await pool.query("INSERT INTO contacts (first_name, last_name, email, phone, address1, address2, city, state, zip, categories, photo_url, photo_filename, photo_mimetype, photo_upload_time, description) VALUES($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15) RETURNING *", [first_name, last_name, email, phone, address1, address2, city, state, zip, categories, photo_url, photo_filename, photo_mimetype, photo_upload_time, description]
+        const user_id = req.user.uid;
+
+        const newContact = await pool.query("INSERT INTO contacts (user_id, first_name, last_name, email, phone, address1, address2, city, state, zip, categories, photo_url, photo_filename, photo_mimetype, photo_upload_time, notes) VALUES($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16) RETURNING *", [user_id, first_name, last_name, email, phone, address1, address2, city, state, zip, categories, photo_url, photo_filename, photo_mimetype, photo_upload_time, notes]
         );
 
         res.status(201).json(newContact.rows[0]);
@@ -68,25 +101,34 @@ app.post("/contacts", parser.single("photo"), async (req, res) => {
 
 // Get all contacts
 
-app.get("/contacts", async (req, res) => {
+app.get("/contacts", verifyToken, async (req, res) => {
     try {
-        const allContacts = await pool.query("SELECT * FROM contacts");
-        res.json(allContacts.rows);
+        const { rows } = await pool.query('SELECT * FROM contacts WHERE user_id = $1', [req.user.uid]);
+        res.send(rows);
     } catch (err) {
         console.error(err.message);
         res.status(500).json({ error: "Server error" });
     }
+    
+    // try {
+    //     const allContacts = await pool.query("SELECT * FROM contacts");
+    //     res.json(allContacts.rows);
+    // } catch (err) {
+    //     console.error(err.message);
+    //     res.status(500).json({ error: "Server error" });
+    // }
 });
 
 // Get a contact
 
-app.get("/contacts/:id", async (req, res) => {
+app.get("/contacts/:id", verifyToken, async (req, res) => {
     try {
         const { id } = req.params;
-        const contact = await pool.query("SELECT * FROM contacts WHERE contacts_id = $1", [id]);
+        const { uid } = req.user; // get uid from decoded token
+        const contact = await pool.query("SELECT * FROM contacts WHERE contacts_id = $1 AND user_id = $2", [id, uid]);
 
         if (contact.rows.length === 0) {
-            return res.status(404).json({ error: 'Contact not found' });
+            return res.status(404).json({ error: 'Contact not found or you do not have persmission to view it' });
         }
 
         res.json(contact.rows[0]);
@@ -98,18 +140,19 @@ app.get("/contacts/:id", async (req, res) => {
 
 // Update a contact
 
-app.put("/contacts/:id", parser.single("photo"), async (req, res) => {
+app.put("/contacts/:id", verifyToken, parser.single("photo"), async (req, res) => {
     try {
         const { id } = req. params;
-        const { first_name, last_name, email, phone, address1, address2, city, state, zip, categories, description } = req.body;
+        const { first_name, last_name, email, phone, address1, address2, city, state, zip, categories, notes } = req.body;
+        const { uid } = req.user;
 
         if (!first_name) {
             return res.status(400).json({ error: 'Missing required field' });
         }
 
-        const contact = await pool.query("SELECT * FROM contacts WHERE contacts_id = $1", [id]);
+        const contact = await pool.query("SELECT * FROM contacts WHERE contacts_id = $1 AND user_id = $2", [id, uid]);
         if (!contact.rows[0]) {
-            return res.status (404).json({ error: 'Contact not found' });
+            return res.status (404).json({ error: 'Contact not found or you do not have permission to update it' });
         }
 
         let { photo_url, photo_filename, photo_mimetype, photo_upload_time } = contact.rows[0];
@@ -121,7 +164,7 @@ app.put("/contacts/:id", parser.single("photo"), async (req, res) => {
             photo_upload_time = new Date();
         }
 
-        const updateContact = await pool.query("UPDATE contacts SET first_name = $1, last_name = $2, email = $3, phone = $4, address1 = $5, address2 = $6, city = $7, state = $8, zip = $9, categories = $10, photo_url = $11, photo_filename = $12, photo_mimetype = $13, photo_upload_time = $14, description = $15 WHERE contacts_id = $16", [first_name, last_name, email, phone, address1, address2, city, state, zip, categories, photo_url, photo_filename, photo_mimetype, photo_upload_time, description, id]);
+        const updateContact = await pool.query("UPDATE contacts SET first_name = $1, last_name = $2, email = $3, phone = $4, address1 = $5, address2 = $6, city = $7, state = $8, zip = $9, categories = $10, photo_url = $11, photo_filename = $12, photo_mimetype = $13, photo_upload_time = $14, notes = $15 WHERE contacts_id = $16 AND user_id = $17", [first_name, last_name, email, phone, address1, address2, city, state, zip, categories, photo_url, photo_filename, photo_mimetype, photo_upload_time, notes, id, uid]);
 
         res.json({ message: "Contact was updated successfully" });
     } catch (err) {
@@ -132,13 +175,14 @@ app.put("/contacts/:id", parser.single("photo"), async (req, res) => {
 
 // Delete a contact
 
-app.delete("/contacts/:id", async (req, res) => {
+app.delete("/contacts/:id", verifyToken, async (req, res) => {
     try {
         const { id } = req.params;
-        const contact = await pool.query("SELECT * FROM contacts WHERE contacts_id = $1", [id]);
+        const { uid } = req.user;
+        const contact = await pool.query("SELECT * FROM contacts WHERE contacts_id = $1 AND user_id = $2", [id, uid]);
         
         if (!contact.rows[0]) {
-            return res.status(404).json({ error: 'Contact not found' });
+            return res.status(404).json({ error: 'Contact not found or you do not have permission to delete it' });
         }
 
         let { photo_url } = contact.rows[0];
@@ -151,7 +195,7 @@ app.delete("/contacts/:id", async (req, res) => {
             await cloudinary.uploader.destroy(public_id);
         }
 
-        const deleteContact = await pool.query("DELETE FROM contacts WHERE contacts_id = $1", [id]);
+        const deleteContact = await pool.query("DELETE FROM contacts WHERE contacts_id = $1 AND user_id = $2", [id, uid]);
 
         res.json({ message: "Contact deleted successfully" });
     } catch (err) {
